requirements.txt
# Core dependencies
PyQt6==6.6.0
PyQt6-Qt6==6.6.0

# AI and ML
torch>=2.2.0
transformers>=4.40.0
accelerate>=0.30.0
bitsandbytes>=0.43.0
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0

# Voice processing
faster-whisper>=1.0.0
TTS>=0.22.0
sounddevice==0.4.6
numpy>=1.22.0
scipy==1.11.4
librosa==0.10.1
pydub==0.25.1

# HTTP and API
requests>=2.32.0
aiohttp>=3.10.0
websockets>=13.0
flask>=3.1.0
flask-socketio>=5.4.0
feedparser==6.0.10
beautifulsoup4==4.12.2
lxml==4.9.3

# Utilities
PyYAML==6.0.1
python-dotenv==1.0.0
pillow==10.1.0
packaging==23.2

# Thai language processing
pythainlp>=5.0.0
thai-segmenter>=0.4.2

# Supabase and Database
supabase==2.3.4
psycopg2-binary==2.9.9

# Computer Vision and Multimodal AI
opencv-python>=4.10.0
transformers[vision]>=4.40.0
timm==0.9.12
clip-by-openai==1.0
easyocr==1.7.0
pytesseract==0.3.10
ultralytics==8.0.225

# Additional Vision Libraries
albumentations==1.3.1
scikit-image==0.22.0
scikit-learn==1.3.2
imageio==2.33.1
av==10.0.0

# Thai OCR specific
thai-segmenter==0.4.2
attrdict==2.0.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0